requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pyodbc==5.1.0
requests==2.32.3
python-dotenv==1.0.1
jinja2==3.1.4
aiohttp==3.10.5
twilio==9.2.3
sentry-sdk[fastapi]
pytz